Introduce the mkfs.sqlitefs utility
The current binary creates the database if it is missing at init() time.
It may be useful to have a dedicated utility to create the file and
initialize the database as the mkfs utilities do.

This introduces the mkfs.sqlitefs utility that creates a SQLiteFS
database in the file named fs.db.

Note: The project produces a single binary and uses a symlink to
sqlitefs, as busybox[1] does.

[1]: https://busybox.net