Implement mknod and utimens callbacks
The callbacks mknod() and utimens() needs to be implemented to be able
to create a new file; along with getattr(), that is implemented already.

The test uses the command touch to create a new empty file that implies
to implement the utimes() callback.

The mknod() callback is called once the file is created.

First, the getattr() callback is invoked to test if the file exists. If
it does not exist, the create() callback is invoked then (or mknod() if
the create() callback is not implemented[*]) to create the file at the
given path. It checks in the end if the file was created by calling
getattr() again.

The utimens() callback is called once the access and modification times
are changed.

The utimens() callback is called in the end to update the time of last
access and last data modification for the file at the given path.

	Running #4: Touch file...
	unique: 22, opcode: LOOKUP (1), nodeid: 1, insize: 48, pid: 25452
	LOOKUP /touched
	getattr /touched
	   unique: 22, error: -2 (No such file or directory), outsize: 16
	unique: 24, opcode: CREATE (35), nodeid: 1, insize: 64, pid: 25452
	create flags: 0x8841 /touched 0100644 umask=0022
	sqlitefs_create(path: /touched, mode: 33188, fi: 0x7f4e3eeb7cf0)
	sqlitefs_create: Function not implemented
	   unique: 24, error: -38 (Function not implemented), outsize: 16
	unique: 26, opcode: MKNOD (8), nodeid: 1, insize: 64, pid: 25452
	create flags: 0xc1 /touched 0100644 umask=0022
	sqlitefs_create(path: /touched, mode: 33188, fi: 0x7f4e3e694bc0)
	sqlitefs_create: Function not implemented
	mknod /touched 0100644 0x0 umask=0022
	getattr /touched
	  File: /touched
	  Size: 0	Blocks: 0	IO Block: 0	regular file
	Device: 0/0	Inode: 0	Links: 1
	Access: (O100644)	Uid: ( 1000/ gportay)	Gid: ( 1000/ gportay)
	Access: 2020-01-03 01:38:47.000000000
	Modify: 2020-01-03 01:38:47.000000000
	Change: 2020-01-03 01:38:47.000000000
	 Birth: -
	   NODEID: 3
	   unique: 26, success, outsize: 144
	unique: 28, opcode: SETATTR (4), nodeid: 3, insize: 128, pid: 25452
	utimens /touched 1578033527.056719879 1578033527.056719879
	getattr /touched
	  File: /touched
	  Size: 0	Blocks: 0	IO Block: 0	regular file
	Device: 0/0	Inode: 0	Links: 1
	Access: (O100644)	Uid: ( 1000/ gportay)	Gid: ( 1000/ gportay)
	Access: 2020-01-03 01:38:47.000000000
	Modify: 2020-01-03 01:38:47.056719879
	Change: 2020-01-03 01:38:47.056719879
	 Birth: -
	   unique: 28, success, outsize: 120
	#4: Touch file: [OK]

[*]: The callback create() will be implemented later.