Implement chmod callback
The chmod() callback is called once the mode of a file is changed.

First, the getattr() callback is invoked to test if the file exists. If
it exists, the chmod() callback is invoked then to update the mode of
the the file at the given path. It checks in the end if the mode was
changed by calling getattr() again.

	Running #6: Change mode...
	unique: 32, opcode: SETATTR (4), nodeid: 3, insize: 128, pid: 32203
	getattr /touched
	  File: /touched
	  Size: 0	Blocks: 0	IO Block: 0	regular file
	Device: 0/0	Inode: 0	Links: 1
	Access: (O100664)	Uid: (    0/    root)	Gid: (    0/    root)
	Access: 2020-01-01 12:02:38.000000000
	Modify: 2020-01-01 12:03:44.326823535
	Change: 2020-01-01 12:03:44.326823535
	 Birth: -
	   unique: 32, success, outsize: 120
	664
	#6: Change mode: [OK]