Implement unlink callback
The unlink() callback is called once the file is removed.

First, the getattr() callback is invoked to test if the file exists. If
it exists, the unlink() callback is invoked then to remove the file at
the given path. It checks in the end if the file was removed by calling
getattr() again.

	Running #11: Remove file...
	unique: 70, opcode: UNLINK (10), nodeid: 1, insize: 50, pid: 33517
	unlink /README.md
	   unique: 70, success, outsize: 16
	unique: 72, opcode: FORGET (2), nodeid: 5, insize: 48, pid: 0
	FORGET 5/1
	DELETE: 5
	unique: 74, opcode: LOOKUP (1), nodeid: 1, insize: 50, pid: 33466
	LOOKUP /README.md
	getattr /README.md
	   unique: 74, error: -2 (No such file or directory), outsize: 16
	#11: Remove file: [OK]