Implement mkdir callback
The mkdir() callback is called once the directory is created.

First, the getattr() callback is invoked to test if the directory
exists. If it does not exists, the mkdir() callback is invoked then to
create the directory at the given path. It checks in the end if the
directory was created by calling getattr() again.

	Running #12: Make directory...
	unique: 84, opcode: LOOKUP (1), nodeid: 1, insize: 44, pid: 34087
	LOOKUP /dir
	getattr /dir
	   unique: 84, error: -2 (No such file or directory), outsize: 16
	unique: 86, opcode: MKDIR (9), nodeid: 1, insize: 52, pid: 34087
	mkdir /dir 0755 umask=0022
	getattr /dir
	  File: /dir
	  Size: 0	Blocks: 0	IO Block: 0	directory
	Device: 0/0	Inode: 0	Links: 2
	Access: (O40755)	Uid: ( 1000/ gportay)	Gid: ( 1000/ gportay)
	Access: 2020-01-01 12:08:59.000000000
	Modify: 2020-01-01 12:08:59.000000000
	Change: 2020-01-01 12:08:59.000000000
	 Birth: -
	   NODEID: 5
	   unique: 86, success, outsize: 144
	#12: Make directory: [OK]