Implement rmdir callback
The rmdir() callback is called once the directory is removed.

First, the getattr() callback is invoked to test if the directory
exists. If it exists, the rmdir() callback is invoked then to remove the
directory to the given path. It checks in the end if the directory was
removed by calling getattr() again.

	Running #13: Remove directory...
	unique: 86, opcode: RMDIR (11), nodeid: 1, insize: 44, pid: 34487
	rmdir /dir
	   unique: 86, success, outsize: 16
	unique: 88, opcode: FORGET (2), nodeid: 5, insize: 48, pid: 0
	FORGET 5/1
	DELETE: 5
	unique: 90, opcode: LOOKUP (1), nodeid: 1, insize: 44, pid: 34434
	LOOKUP /dir
	getattr /dir
	   unique: 90, error: -2 (No such file or directory), outsize: 16
	#13: Remove directory: [OK]