Implement symlink callback
The symlink() callback is called once the symbolic link is created.

First, the getattr() callback is invoked to test if the symbolic link
exists. If it does not exists, the symlink() callback is invoked then to
create the symbolic link at the given path. It checks in the end if the
symbolic link was created by calling getattr() again.

	Running #14: Make symbolic link...
	unique: 94, opcode: LOOKUP (1), nodeid: 1, insize: 48, pid: 34996
	LOOKUP /symlink
	getattr /symlink
	   unique: 94, error: -2 (No such file or directory), outsize: 16
	unique: 96, opcode: SYMLINK (6), nodeid: 1, insize: 55, pid: 34996
	symlink .Trash /symlink
	getattr /symlink
	  File: /symlink
	  Size: 0	Blocks: 0	IO Block: 0	symbolic link
	Device: 0/0	Inode: 0	Links: 2
	Access: (O120000)	Uid: ( 1000/ gportay)	Gid: ( 1000/ gportay)
	Access: 2020-01-01 12:11:12.000000000
	Modify: 2020-01-01 12:11:12.000000000
	Change: 2020-01-01 12:11:12.000000000
	 Birth: -
	   NODEID: 6
	   unique: 96, success, outsize: 144
	#14: Make symbolic link: [OK]