readdir_cb: fix sqlite3 callback
There is a *BIG* misunderstanding of how is invocked the callback given
to sqlite3_exec().

Regarding the documentation of sqlite3_exec()[1]:

	If the callback function of the 3rd argument to sqlite3_exec()
	is not NULL, then it is invoked for each result row coming out
	of the evaluated SQL statements.

That is to say, it is wrong to iterate over the argv[] arguments.

Anyway, the argc returned by sqlite3_exec() is always 1 as there is one
single column selected by the SQL request.

[1]: https://www.sqlite.org/c3ref/exec.html