add_{file|directory|symlink}: deduce parent from path
The parent is root for every files (i.e. /). It is hardcoded since
always. As a consequence, sqlitefs is able to manage up to one path
component; that is a very bad limitation.

This commit deduces the parent from the path and allows to manage a
unlimited[*] number of path component.

[*]: limited by the path length MAX_PATH or SQLITE_MAX_LENGTH[1].
[1]: https://www.sqlite.org/limits.html