readdir_cb: fix none-top-level path component
The trick to remove the parent component using the length of the parent
works for the files under the top-level directory only (i.e. the root
directory /).

The filler is filled with:

	/ -> 1
	-------
	/a -> a
	/b -> b

For the files that are located in sub-directories, the filler is filled
with:

	/a -> 2
	-------
	/dir/a -> /a
	/dir/b -> /b

That logic is wrong an causes libfuse to return -EIO because of the
leading '/' in the file.

The trick does not work for element with additional directories (i.e.
none-top-level directory): it must be length + 1 if the length is 1
(i.e. parent is root /).

	len = strlen(pdata->parent);
	if (len == 1)
		len++;

Lets use the function basename() instead of doing the little hack above.

Fixes
	$ make shell
	$ mkdir mountpoint/dir
	$ touch mountpoint/dir/file
	$ ls mountpoint/dir/
	find: ‘mountpoint/dir’: Input/output error