Detect orphans and print them at startup
The orphans files are inodes that have lost its directory parent. They
cannot be reached anymore via a mounted file-system.

It is easy to create orphans as the function __rmdir() does not checks
for the directory emptiness before removing it.

The man page says:

	The rmdir() function shall remove a directory whose name is
	given by path. The directory shall be removed only if it is an
	empty directory.

This commit outputs the list of the orphans on startup to warn the user.
For now, nothing can be done to reach or remove them.