Replace .Trash by .lost+found
The directory .Trash is around since always to populate the initial
file-system.

This directory is now useless since sqlitefs can create its own files
and directories when the database is mounted.

This commit replaces the directory .Trash by the directory .lost+found
that is going to hold recovered files when bad things happen.