Introduce the fsck.sqlitefs utility
As for the mkfs.sqlitefs utility introduced in commit 48759f8 (Introduce
the mkfs.sqlitefs utility), this commit introduces the fsck.sqlitefs
utility to deal with orphans.

It re-uses the support script sqlitefs-fsck as it was easier to write a
script to execute the SQLite requests in bash than to develop the same
feature in C.

Note: The project produces a single binary and uses a symlink to
sqlitefs, as busybox[1] does.

[1]: https://busybox.net