Fix the calls to system()
The calls to system() return positive value if the command that is run
exits with a non-zero exit code.

As a consequence, it outputs:

	system: Success.

The function perror() must be invoked only if the function system() returns
-1.

Moreover, the man page of wait(3p) says:

	WIFEXITED(stat_val)

		Evaluates to a non-zero value if status was returned for
		a child process that terminated normally.

	WEXITSTATUS(stat_val)

		If the value of WIFEXITED(stat_val) is non-zero, this
		macro evaluates to the low-order 8 bits of the status
		argument  that the  child  process  passed  to _exit()
		or exit(), or the value the child process returned from
		main().

The real exit code of the command is returned by the macro WEXITSTATUS;
this macro must be run if WIFEXITED as returned a non-zero value (i.e.
if the process has exited normally, not signaled).