Import and modify the code of the function fuse_main_real()
The function fuse_main()[1] is a ready-to-use entry-point for any new
FUSE module.

	This is for the lazy. This is all that has to be called from the main() function.

This function is not versatile enough for the positional parameters.

The idea is to let the path to the sqlite database file configurable.
The order of the positional parameters matters. The path to the file
must be the first argument and the mountpoint the second; as for the
mount utility.

	mount [-t fstype] [-o options] device dir

The utility fusermount3, and the function fuse_main() besides, take the
mountpoint as the single positional argument.

	fusermount3 [OPTIONS] MOUNTPOINT

This commit copies the code of the entry-point fuse_main_real()[2], the
function fuse_parse_cmdline()[3], the helper fuse_helper_opt_proc[3] and
the necessary macro and structure[5] that are hidden in the source file
lib/helper.c[6].

This provides a good start of skeleton to rewrite it to the needs.

[1]: http://libfuse.github.io/doxygen/include_2fuse_8h.html#ac99b844cee7aaa8fb4e35df5b5488d82
[2]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/helper.c#L279-L358
[3]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/helper.c#L202-L222
[4]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/helper.c#L142-L172
[5]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/helper.c#L29-L54
[6]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/helper.c