chdir() back to $PWD
Since the previous commit 16c85d8 (Import and modify the code of the
function fuse_main_real()), the current working directory is the root
directory.

	$ EXEC=/bin/sh sqlitefs -f mountpoint
	sh-5.0$ ls
	bin  boot  dev	etc  home  lib	lib64  media  mnt  opt	proc  root  run  sbin  srv  swapfile  sys  tmp	usr  var
	sh-5.0$

Before that commit, the command was started very early. Now, it is
started after the file-system is mounted and after the process is
daemonized (or not).

The function fuse_daemonize() chdir to root directory since the commit
c52c325[1] (fuse_daemonize(): chdir to "/" even if not running in the
background).

This commit chdir() back to $PWD before the command is run.

[1]: https://github.com/libfuse/libfuse/commit/c52c3256a5b0a586ab0c529ef76505143523780a