Add Docker image
This commits adds a Dockerfile and the docker-shell Makefile target that
build a docker image with unstripped version of both sqlite and fuse
libraries. This image is suitable for development.

The Dockerfile creates an image with the development tools installed,
sets up the current user, compiles both sqlite and fuse with the debug
symbols and installs them to the root. It installs strace and valgrind
as well.

The docker-shell Makefile target sets the user rights from the host,
bind-mounts the sources and opens the image.