Add dosh support
Dosh is a shell frontend for docker that builds and runs a container
using the current user and binds mount the current working directory.

It is a bash scripts that wraps the calls to docker-build and
docker-run.

This commit adds the doshrc file to specify additionnal arguments to
build the image and run sqlitefs in the container.

[1]: https://www.portay.io/dosh/