Allows to specify the command to run in foreground
Setting the environment variable EXEC is not very convenient and it is
hard to not to break the list of argument (when somes contain spaces).

The command to run, as well its arguments, are appended to the full
command line, right after the mount-point. It replaces and deprecates
the environment variable EXEC. See the new usage:

	usage: sqlitefs [options] <file> <mountpoint> [--] [command] [args]

The example below opens an interective shell:

	$ sqlitefs fs.db moutpoint /bin/sh
	sh-5.0$

As FUSE handles the double dashes `--', the following command is also
valid:

	$ sqlitefs fs.db mountpoint -- /bin/sh
	sh-5.0$

Note: the command implies the foreground mode.