sqlitefs_main: return 0 if SIGTERM received
The functions fuse_loop()[1] and fuse_loop_mt() returns the session
error[1]. The session error is the set to the signal number that is
received in the signal handler exit_handler()[3].

The exit_handler() is set for the signals SIGUP, SIGINT and SIGTERM[3];
the SIGPIPE is set to the handler do_nothing() that does nothing.

This commit forces the function sqlitefs_main() to return 0 if the
function fuse_loop[_mt]() returns SIGTERM (i.e. the loop was interrupted
by the signal SIGTERM, when the command exits).

Note: Both functions fuse_loop() and fuse_loop_mt() return 0 if the
mountpoint is unmount using fusermount.

[1]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_loop.c#L42-L43
[2]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_loop_mt.c#L348-L349
[3]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_signals.c#L30
[4]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_signals.c#L69-L71