sqlitefs_main: return the exit status and signal
Both functions fuse_loop() and fuse_loop_mt() returns either 0, or a
negative errno value, or the signal number if it is signaled.

The function fuse_loop() has two cases if the remember or noforget
option is set:
	- the function fuse_session_loop_remember() returns either 0, or
	  -1[1].
	- the function fuse_session_loop() returns either 0[2], or the
	  negative errno value returned by the function
	  fuse_session_receive_buf_int()[3][4][1][*], or the signal
	  number[4] returned by the session error[5] set by the
	  exit_handler()[6].

The function fuse_loop_mt() returns either 0[7], or the negative errno
value returned by the function fuse_session_receive_buf_int()[8][9][10],
or the signal number returned by the session error[11] set by the
exit_handler()[6].

Thus, the only positive values are the signal number that have
interrupted the fuse_loop[_mt]() functions.

This commit makes the function sqlitefs_main() return the process status
as if it was returned by wait.

The value that is returned by the function sqlitefs_main() must be
interpreted using macros defined in the header <sys/wait.h> (i.e.
WIFEXITED(), WEXITSTATUS(), WIFSIGNALED() and WTERMSIG()).

[1]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse.c#L4558
[2]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_loop.c#L38-L41
[3]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_lowlevel.c#L2736
[4]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse.c#L4541
[5]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_loop.c#L42-L43
[6]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_signals.c#L30
[7]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_loop_mt.c#L317
[8]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_loop_mt.c#L343
[9]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_loop_mt.c#L134
[10]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_loop_mt.c#L127
[11]: https://github.com/libfuse/libfuse/blob/fuse-3.9.0/lib/fuse_loop_mt.c#L348-L349
[*]: A positive value returned by fuse_session_receive_buf_int()[3] is
returned as zero[1].