Move the exit status logic in sqlitefs_main()
Since the commit 424fba9 (Run a command in foreground mode), the
function main() does not return the result of the function
fuse_main() directly.

Then, the part of code that handles the command moved in the function
sqlitefs_main() (a modified version of the FUSE helper fuse_main_real())
since commit 16c85d8 (Import and modify the code of the function
fuse_main_real()).

Finally, the commit d84eab2 (sqlitefs_main: return the exit status and
signal) makes the function sqlitefs_main() returns an exit status that
must be interpreted using the macros WIFEXITED(), WEXITSTATUS(),
WIFSIGNALED() and WTERMSIG(); that is weird.

This commit moves the logic of the exit status in the function
sqlitefs_main(). It removes the necessity of interpreting the returned
value using the Wxxx() macros in the function main() and simplifies it.