rmdir: do not remove non-empty directory
The actual implementation of the rmdir() callback does not check for its
emptiness before removing the directory. The children of the directory
become orphans.

The man-page of rmdir says:

	If the directory is not an empty directory, rmdir() shall fail
	and set errno to [EEXIST] or [ENOTEMPTY].

This commit reuses the code used to implement the readdir() callback.
It is moved to a the new function __readdir(); and it returns the number
of children the directory has.

Both callbacks readdir() and rmdir() are rewritten on top of that new
function. The rmdir() callback returns ENOEMPTY if the path is not empty
and does not remove the directory.