Use the rowid to represent the st_ino attribute
The attribute st_ino is never used. Its value is zero and stored as is.

This commit uses the rowid instead. The rowid is a good candidate to
represent the st_ino as it is unique.

Note that libfuse and the kernel use an internal ino (that is also
called the nodeid). See the use_ino[1].

[1]: https://libfuse.github.io/doxygen/structfuse__config.html#adf78e64e79e31c4fe1464cd0744ea725