fprintstat: introduce TIMSIZ macro
The macro BUFSIZ depends on the architecture and it is set from 2k up to
8k (and more?).

Allocating many buffers of that size in the stack is terrible and may
causes a stack overflow.

The macro TIMSIZ defines a relatively sane buffer size to store the
file mode type string (32 bytes).

This commit saves some allocations performed in the stack by reducing
the size of a stack allocated buffer to 32 bytes only (was previously
several kilo-bytes).