Introduce the option verbose
The SQLiteFS outputs for debug are poluted by the too many FUSE debug
outputs.

This commit introduces the new option verbose to make the distinction
between some useful SQLiteFS outputs that are enabled using the option
--verbose and some debug outputs that are enabled using the FUSE option
--debug.

For now, the option verbose outputs the stat of the file when it is
requested by FUSE. The read and write buffer are output if the option
debug is set twice.

Note: the Makefile introduces the verbose-shell target in the same time.