timespec_r: output the numeric timezone
The stat command outputs the numeric timezone for the last access
timestamp, the last modification timestamp, the last status change
timestamp and the file creation timestamp (birth).

The function fprintstat() does not output that information via the
function timespec_r().

The man page for function strftime() says:

	%z     The +hhmm or -hhmm numeric timezone (that is, the hour
	and minute offset from UTC). (SU)

This commit outputs that information that is given by the %z placeholder
in the function strftime(). The idea is to implement the function
fprintstat() as close as possible as the output of the stat command.