fprintstat: fix the access mode output
The function fprintstat() outputs the full mode value where as the stat
command outputs the file permissions onlys. See both output below:

The output of the fprintstat() function:

	  File: /symlink
	  Size: 6	Blocks: 0	IO Block: 0	symbolic link
	Device: 0/0	Inode: 3	Links: 2
	Access: (120777/lrwxrwxrwx)	Uid: ( 1000/ gportay)	Gid: ( 1000/ gportay)
	Access: 2020-01-22 04:19:00.299513371 -0500
	Modify: 2020-01-22 04:19:00.299513371 -0500
	Change: 2020-01-22 04:19:00.299513371 -0500
	 Birth: -

The output of the stat command:

	  File: mountpoint/symlink -> .Trash
	  Size: 6         	Blocks: 0          IO Block: 4096   symbolic link
	Device: 3fh/63d	Inode: 3           Links: 2
	Access: (0777/lrwxrwxrwx)  Uid: ( 1000/ gportay)   Gid: ( 1000/ gportay)
	Access: 2020-01-22 04:19:00.299513371 -0500
	Modify: 2020-01-22 04:19:00.299513371 -0500
	Change: 2020-01-22 04:19:00.299513371 -0500
	 Birth: -

This commits does a bitwise AND operation with the mask ~S_IFMT to
remove the bits field corresponding to the file type.