Use the function execvp()
The version `p' of the function execv() uses the environment variable
PATH to run the command. In other words, it is not necessary to give the
full path to the executable.

This commit replaces the use of the function execv() by execvp(). It is
not necessary to give the full path to the command anymore; and it works
as before the commit 5d61e20 (Allows to specify the command to run in
foreground).